API: Move caproto._cli to caproto.sync.client.
