+++ scripts/caproto-get
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-from caproto._cli import get_cli
+from caproto.sync.client import get_cli
 
 
 if __name__ == '__main__':
